requirements.txt
anthropic
google-api-python-client
google-auth-oauthlib
youtube-transcript-api
yt-dlp
python-dotenv
click
tzdata
